requirements.txt
anyio==3.6.1
asgiref==3.5.2
attrs==22.1.0
bleach==5.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-log==0.4.0
colorama==0.4.5
cryptography==37.0.4
deprecation==2.1.0
Django==3.2
django-cors-headers==3.7.0
django-extensions==3.1.3
django-filter==2.4.0
django-injector==0.2.2
django-model-utils==4.1.1
django-redis==4.12.1
djangorestframework==3.12.4
docutils==0.19
dotty-dict==1.3.1
gitdb==4.0.9
GitPython==3.1.27
gotrue==0.5.0
gunicorn==20.1.0
h11==0.12.0
httpcore==0.14.7
httpx==0.21.3
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
injector==0.20.1
invoke==1.7.1
jeepney==0.8.0
keyring==23.8.2
packaging==21.3
pkginfo==1.8.3
pluggy==0.13.1
postgrest-py==0.10.2
py==1.11.0
pycparser==2.21
pydantic==1.9.2
Pygments==2.13.0
pyparsing==3.0.9
pytest==6.2.3
python-dateutil==2.8.1
python-dotenv==0.17.1
python-gitlab==3.8.1
python-semantic-release==7.28.1
pytz==2022.2.1
readme-renderer==36.0
realtime==0.0.4
redis==3.5.3
requests==2.28.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
SecretStorage==3.3.3
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
sqlparse==0.4.2
storage3==0.3.4
supabase==0.5.8
toml==0.10.2
tomlkit==0.10.2
tqdm==4.64.0
twine==3.8.0
typing_extensions==4.3.0
urllib3==1.26.11
webencodings==0.5.1
websockets==9.1
zipp==3.8.1
